Finally finished the dev branch
