--- asherTest.js
@@ -1,5 +1,0 @@
-const brain = require(process.cwd() + `/core/functions/latestBrain`);
-
-const Asher = new brain();
-Asher.loadNeurons();
-Asher.loadAxons();
