--- core/app/brain_associations.js
@@ -1,8 +1,0 @@
-var mongoose = require("mongoose");
-var Schema = mongoose.Schema;
-
-var brainAssociations = new Schema({
-	word_id: String,
-	sentance_id: String,
-	weight: String,
-});
