--- core/app/brain_sentences.js
@@ -1,7 +1,0 @@
-var mongoose = require("mongoose");
-var Schema = mongoose.Schema;
-
-var brainSentences = new Schema({
-	sentance: String,
-	used: Number,
-});
