--- core/app/brain_words.js
@@ -1,6 +1,0 @@
-var mongoose = require("mongoose");
-var Schema = mongoose.Schema;
-
-var brainWords = new Schema({
-	word: String,
-});
