+++ core/functions/brain.js
@@ -1,4 +1,10 @@
 /*
+    THIS VERSION IS NO LONGER IN USE... IT IS HERE
+    TO SALVAGE PARTS FROM WHILST THE NEW VERSION IS
+    BEING MADE...
+*/
+
+/*
     THIS 'BRAIN' IS THE CONVERSATION MODEL
     FOR THE WHOLE OF ASHER. IF THE USER IS
     HAVING A GENERAL CONVERSATION WITH HIM
