+++ mods/greetings/mod.js
@@ -1,7 +1,9 @@
 module.exports.core = (core) => {
     core.Asher.hear(['Good? night'], (payload, chat) => {
-        if (core.monitor.helper.getTime() == core.monitor.helper.timeOfDay.night) {
+        if (core.helper.getTime() == core.helper.timeOfDay.night) {
             chat.say(`Good night`);
+        } else {
+            chat.say(`Nigga... u confused?`)
         }
     });
 }
