--- test.html
@@ -1,103 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Speech Detection</title>
-</head>
-<body>
-
-    <div class="button-wrapper" on-tap="regActive()" onclick="regActive()">
-        <div class="button">
-            <svg class="button-svg" viewBox="0 0 24 24">
-                <path d="M12 15c1.66 0 2.99-1.34 2.99-3l.01-6c0-1.66-1.34-3-3-3s-3 1.34-3 3v6c0 1.66 1.34 3 3 3zm5.3-3c0 3-2.54 5.1-5.3 5.1s-5.3-2.1-5.3-5.1h-1.7c0 3.42 2.72 6.23 6 6.72v3.28h2v-3.28c3.28-.48 6-3.3 6-6.72h-1.7z"
-                />
-                <path d="M0 0h24v24h-24z" fill="none" />
-            </svg>
-        </div>
-    </div>
-
-
-<script>
-    var isListening = false;
-    const btn = document.querySelector(`.button`);
-    var recognition;
-    var result = "";
-
-
-    const regActive = function() {
-        isListening = isListening ? false : true;
-        if (isListening) rec();
-        if (!isListening) recognition.stop();
-    }
-
-    const rec = function() {
-
-        recognition = new webkitSpeechRecognition();
-
-        recognition.continuous = false;
-        recognition.interimResults = false;
-
-        recognition.lang = "en-US";
-        recognition.start();
-        btn.classList.toggle("active");
-        recognition.onresult = function (e) {
-            result = e.results[0][0].transcript;
-            console.log(result);
-            if (btn.classList.contains('active')) {
-                document.querySelector(`.button`).classList.toggle('active');
-            }
-            if (isListening) {
-                // This means the user just let it run out...
-                isListening = isListening ? false : true;
-            }
-            recognition.stop();
-        };
-
-        recognition.onerror = function (e) {
-            recognition.stop();
-        }
-
-    }
-
-
-  
-</script>
-
-
-<style>
-    .button {
-        will-change: transform;
-        width: 140px;
-        height: 140px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        -webkit-transform: translate(-50%, -50%);
-                transform: translate(-50%, -50%);
-        background: #5fb5fe;
-        border-radius: 50%;
-        -webkit-transform-origin: left top;
-                transform-origin: left top;
-        transition: all 0.6s cubic-bezier(0.23, 1, 0.32, 1);
-        cursor: pointer;
-    }
-    .active {
-        background: #ff0000;
-    }
-    .button-svg {
-        width: 80px;
-        height: 80px;
-        fill: white;
-        transition: all 0.6s cubic-bezier(0.23, 1, 0.32, 1);
-    }
-    .button:hover svg {
-        fill: rgba(0, 0, 0, 0.2);
-    }
-
-</style>
-
-</body>
-</html>
